Remove admin menu. Remove some settings from admin settings. Add modifiers enum and handling. Cleanup card data evaluation. Some fixes from PlatformIO.s

+++ Tonuino_Main/src/shared/Tonuino_Player.h
@@ -23,7 +23,7 @@ class TonuinoTimer
 {
 	public:
 	static unsigned long activeTime;
-	static long timeInMin;
+	static uint8_t timeInMin;
 	
 	void activate();
 	void disable();
